requirements.txt
numpy
matplotlib>=2.1.0
pandas>= 0.22.0
scikit-learn>=0.19.1
scipy>=1.0.0
svgwrite>=1.1.12
tensorflow==2.12.0
tensorflow-probability==0.20.1
cairosvg
pillow
cx_Freeze